refactor custom queries for restaurants

+++ src/main/java/club/beingsoft/restaurants/to/RestaurantWithVotesTo.java
@@ -2,11 +2,9 @@ package club.beingsoft.restaurants.to;
 
 import lombok.Value;
 
-import java.math.BigInteger;
-
 @Value
 public class RestaurantWithVotesTo {
     Integer id;
     String name;
-    BigInteger vcount;
+    Long vcount;
 }
